+++ modules/webgl/src/context/debug/spector.ts
@@ -7,6 +7,7 @@ import {loadScript} from '../../utils/load-script';
 import {getWebGLContextData} from '../helpers/webgl-context-data';
 
 import type {Spector} from './spector-types';
+import {registerSpectorJS} from './debug-hooks';
 
 /** Spector debug initialization options */
 type SpectorProps = {
@@ -106,3 +107,5 @@ export function initializeSpectorJS(props: SpectorProps): Spector | null {
 
   return spector;
 }
+
+registerSpectorJS({load: loadSpectorJS, initialize: initializeSpectorJS});