+++ modules/webgl/src/debug.ts
@@ -1,0 +1,9 @@
+// luma.gl
+// SPDX-License-Identifier: MIT
+// Copyright (c) vis.gl contributors
+
+export {
+  loadWebGLDeveloperTools,
+  makeDebugContext
+} from './context/debug/webgl-developer-tools';
+export {loadSpectorJS, initializeSpectorJS} from './context/debug/spector';